# api/openapi.yaml
openapi: 3.0.3
info:
  title: Warrant API
  description: REST API for Warrant ticketing (orgs, projects, tickets, reviews, queue).
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Local
paths:
...
  /orgs:
    get:
      operationId: ListOrgs
      summary: List organizations for the authenticated user (OAuth required)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Org'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
    post:
      operationId: CreateOrg
      summary: Create organization (auth required)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrgRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Org'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
...
  /orgs/{orgID}/projects:
    get:
      operationId: ListProjectsByOrg
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          description: Filter by project status; default active.
          schema:
            type: string
            enum: [active, closed, all]
            default: active
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      operationId: CreateProject
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
  /projects/{projectID}:
    get:
      operationId: GetProject
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
    patch:
      operationId: UpdateProject
      summary: Update project (e.g. set status to active or closed)
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        "400":
          description: Bad request (e.g. invalid status)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
...
  /orgs/{orgID}/projects/{projectID}/git-notes/commits/{commitSha}:
    get:
      operationId: GetCommitNotes
      summary: Get git notes for a commit (requires repo_path query; server returns 501 if repo not accessible)
      security:
        - BearerAuth: []
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
        - name: projectID
          in: path
          required: true
          schema:
            type: string
        - name: commitSha
          in: path
          required: true
          schema:
            type: string
        - name: repo_path
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          description: Note type (decision, trace, or intent); omit for all types
          schema:
            type: string
            enum: [decision, trace, intent]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitNotesResponse'
        "404":
          description: Not found (project or org mismatch)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "501":
          description: Not implemented (repo_path missing or not a git repo)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotImplementedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    StructuredError:
      type: object
      required: [error, code, retriable]
      properties:
        error:
          type: string
        code:
          type: string
          enum: [lease_expired, unauthorized, not_found, conflict, invalid_input, forbidden, internal, not_implemented, project_closed]
        retriable:
          type: boolean
    NotImplementedError:
      type: object
      description: Returned when server cannot access repo (e.g. git-notes without repo_path)
      properties:
        error:
          type: string
        code:
          type: string
          enum: [not_implemented]
...
    CommitNotesResponse:
      type: object
      properties:
        commit_sha:
          type: string
        notes:
          type: object
          additionalProperties:
            type: string
...
    Org:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        created_at:
          type: string
          format: date-time
    CreateOrgRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
    Project:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        name:
          type: string
        slug:
          type: string
        repo_url:
          type: string
        tech_stack:
          type: array
          items:
            type: string
        context_pack:
          type: object
        status:
          type: string
          enum: [active, closed]
          description: active (default) or closed; list endpoints default to active only.
        created_at:
          type: string
          format: date-time
    UpdateProjectRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, closed]
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        repo_url:
          type: string
        tech_stack:
          type: array
          items:
            type: string
...